ubah tipe data annual_revenue
ubah tipe data annual_revenue from integer to bigInteger

+++ database/migrations/2022_09_07_094124_create_companies_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->string('city')->nullable();
             $table->string('postal_code')->nullable();
             $table->integer('number_of_employees')->nullable();
-            $table->integer('annual_revenue')->nullable();
+            $table->bigInteger('annual_revenue')->nullable();
             $table->string('time_zone')->nullable();
             $table->text('description')->nullable();
             $table->string('linkedin_company')->nullable();
